Better header generation for recording scripts
Detects the number of targets or vrcrickets to record, and automatically generate the header. records the arena corner coordinates and the centroid coordinates of any obstacles in the scene. For quality of life improvement, now detects if a suffix.txt file already exists, and if so, saves a copy of it before generating a clean suffix.txt file.