Real cricket tracking with Motive
Created a GameObject to represent the real cricket. Wrote Recorder_script_AR_cricket.cs for tracking and saving the coordinate of the real cricket. This is done by tracking the position of any point coming from Motive that is not assigned to a Rigid body (I assume this to only be 1).